requirements.txt
Flask==2.2.5
python-dotenv==1.0.1
google-api-python-client==2.126.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
presidio-analyzer==2.2.35
presidio-anonymizer==2.2.35
spacy==3.7.2
PyPDF2==3.0.1
python-docx==1.1.0
